refactored client config builder into a base abstract class and interface
