+++ packages/js/client-config-builder/src/index.ts
@@ -1,3 +1,5 @@
-export * from "./ClientConfigBuilder";
-export * from "./ClientConfig";
-export * from "./bundles";
+export * from './BaseClientConfigBuilder';
+export * from './ClientConfig';
+export * from './ClientConfigBuilder';
+export * from './IClientConfigBuilder';
+export * from './bundles';
